Refactor and cleanups
