+++ src/utils/state-management-utils.ts
@@ -19,29 +19,44 @@ export const findMatchingMessageFromMessageList = (messageToMatch: Message, mess
       message.authorRole === messageToMatch.authorRole,
   );
 
+const stateChangingEventMessages = [
+  CHAT_EVENTS.GREETING,
+  CHAT_EVENTS.ASK_PERMISSION_IGNORED,
+  CHAT_EVENTS.ANSWERED,
+  CHAT_EVENTS.TERMINATED,
+  CHAT_EVENTS.UNAVAILABLE_ORGANIZATION,
+  CHAT_EVENTS.UNAVAILABLE_CSAS,
+  CHAT_EVENTS.UNAVAILABLE_HOLIDAY,
+  TERMINATE_STATUS.CLIENT_LEFT_WITH_ACCEPTED,
+  TERMINATE_STATUS.CLIENT_LEFT_WITH_NO_RESOLUTION ,
+  TERMINATE_STATUS.CLIENT_LEFT_FOR_UNKNOWN_REASONS ,
+  TERMINATE_STATUS.ACCEPTED,
+  TERMINATE_STATUS.HATE_SPEECH ,
+  TERMINATE_STATUS.OTHER ,
+  TERMINATE_STATUS.RESPONSE_SENT_TO_CLIENT_EMAIL,
+];
+
 export const isStateChangingEventMessage = (msg: Message): boolean => {
   if(!msg.event)
     return false;
 
-  const whitelist = [
-    CHAT_EVENTS.GREETING,
-    CHAT_EVENTS.ASK_PERMISSION_IGNORED,
-    CHAT_EVENTS.ANSWERED,
-    CHAT_EVENTS.TERMINATED,
-    CHAT_EVENTS.UNAVAILABLE_ORGANIZATION,
-    CHAT_EVENTS.UNAVAILABLE_CSAS,
-    CHAT_EVENTS.UNAVAILABLE_HOLIDAY,
-    TERMINATE_STATUS.CLIENT_LEFT_WITH_ACCEPTED,
-    TERMINATE_STATUS.CLIENT_LEFT_WITH_NO_RESOLUTION ,
-    TERMINATE_STATUS.CLIENT_LEFT_FOR_UNKNOWN_REASONS ,
-    TERMINATE_STATUS.ACCEPTED,
-    TERMINATE_STATUS.HATE_SPEECH ,
-    TERMINATE_STATUS.OTHER ,
-    TERMINATE_STATUS.RESPONSE_SENT_TO_CLIENT_EMAIL,
-  ];
-
   const event = msg.event as CHAT_EVENTS | TERMINATE_STATUS;
-  return whitelist.includes(event) || (event === CHAT_EVENTS.CONTACT_INFORMATION && msg.content?.length === 0);
+  return stateChangingEventMessages.includes(event) || (event === CHAT_EVENTS.CONTACT_INFORMATION && msg.content?.length === 0);
+}
+
+const nonDisplayableEvent = [
+  CHAT_EVENTS.GREETING.toString(),
+  TERMINATE_STATUS.CLIENT_LEFT_WITH_ACCEPTED.toString(), 
+  TERMINATE_STATUS.CLIENT_LEFT_WITH_NO_RESOLUTION.toString(), 
+  TERMINATE_STATUS.CLIENT_LEFT_FOR_UNKNOWN_REASONS.toString(), 
+  TERMINATE_STATUS.ACCEPTED.toString(), 
+  TERMINATE_STATUS.HATE_SPEECH.toString(), 
+  TERMINATE_STATUS.OTHER.toString(), 
+  TERMINATE_STATUS.RESPONSE_SENT_TO_CLIENT_EMAIL.toString(),
+]
+
+export const isDisplayableMessages = (msg: Message): boolean => {
+  return !nonDisplayableEvent.includes(msg.event!);
 }
 
 export const clearStateVariablesFromSessionStorage = (): void => {
